requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
backports.tarfile==1.2.0
Brotli==1.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
duckduckgo_search==5.3.1b1
exceptiongroup==1.2.2
gTTS==2.5.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.8
jaraco.context==6.0.1
Mako==1.3.6
MarkupSafe==3.0.2
more-itertools==10.4.0
openai==1.29.0
pillow==10.3.0
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
python-dotenv==1.0.1
python-telegram-bot==21.1.1
pytube==15.0.0
redis==5.0.8
regex==2024.7.24
requests==2.31.0
six==1.16.0
sniffio==1.3.1
socksio==1.0.0
spotipy==2.23.0
SQLAlchemy==2.0.36
tenacity==8.3.0
tiktoken==0.7.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
whois==0.9.27
wolframalpha==5.0.1
xmltodict==0.13.0
